feat(endpoint-micropub): update database before content store
