+++ packages/endpoint-micropub/tests/unit/post-data.js
@@ -37,6 +37,8 @@ test.beforeEach((t) => {
             };
           }
         },
+        async insertOne() {},
+        async replaceOne() {},
       },
     },
     url: "https://website.example/foo",
